v0.8.29-alpha (Dev_5)
- Fixed video closing X missing/being missaligned.
- Added the messages feature.
- Fixed a bug that showed the wrong button when using the "Anonymous" Discord Activity.
- Added a check for the last searched players feature (checks for file before loading).
- Turning off the rich presence will now do so instantly without having to restart.
- Removed BattlePass Navitem. (Still in production)
- Added a check for the new "message_data" folder on startup.

+++ assets/js/playersearch/playerSearch.js
@@ -1,34 +1,37 @@
 const ipc = require('electron').ipcRenderer;
 const fs = require('fs')
 
-var lastSearchedMatches = fs.readFileSync(process.env.APPDATA + '/VALTracker/user_data/player_search/last_searched_players.json');
-var parsedLastSearchedMatches = JSON.parse(lastSearchedMatches)
-async function deleteSearchedPlayer(DOMElement, playerSlot) {
-  $(DOMElement).remove();
-  $('.removeLastSearch').addClass('disabled');
-  parsedLastSearchedMatches.splice(playerSlot, 1);
-  fs.writeFile(process.env.APPDATA + '/VALTracker/user_data/player_search/last_searched_players.json', JSON.stringify(parsedLastSearchedMatches), err => {
-    if(err) console.log(err)
-    $('.removeLastSearch').removeClass('disabled');
-  })
-}
+if(fs.existsSync(process.env.APPDATA + '/VALTracker/user_data/player_search/last_searched_players.json')) {
+  var lastSearchedMatches = fs.readFileSync(process.env.APPDATA + '/VALTracker/user_data/player_search/last_searched_players.json');
+  var parsedLastSearchedMatches = JSON.parse(lastSearchedMatches)
 
-async function openSearchedPlayer(playerData) {
-  var splitPlayerData = playerData.split(";")
-  var p_uuid = splitPlayerData[0]
-  var p_name = splitPlayerData[1]
-  var p_tag = splitPlayerData[2]
-  var p_reg = splitPlayerData[3]
-
-  sessionStorage.setItem("puuid", p_uuid);
-  sessionStorage.setItem("player_name", p_name);
-  sessionStorage.setItem("player_tag", p_tag);
-  sessionStorage.setItem("player_region", p_reg);
-
-  var page = window.location.href;
-  sessionStorage.setItem("last_page", page);
+  async function deleteSearchedPlayer(DOMElement, playerSlot) {
+    $(DOMElement).remove();
+    $('.removeLastSearch').addClass('disabled');
+    parsedLastSearchedMatches.splice(playerSlot, 1);
+    fs.writeFile(process.env.APPDATA + '/VALTracker/user_data/player_search/last_searched_players.json', JSON.stringify(parsedLastSearchedMatches), err => {
+      if(err) console.log(err)
+      $('.removeLastSearch').removeClass('disabled');
+    })
+  }
   
-  window.location.href = "playerProfile.html";
+  async function openSearchedPlayer(playerData) {
+    var splitPlayerData = playerData.split(";")
+    var p_uuid = splitPlayerData[0]
+    var p_name = splitPlayerData[1]
+    var p_tag = splitPlayerData[2]
+    var p_reg = splitPlayerData[3]
+  
+    sessionStorage.setItem("puuid", p_uuid);
+    sessionStorage.setItem("player_name", p_name);
+    sessionStorage.setItem("player_tag", p_tag);
+    sessionStorage.setItem("player_region", p_reg);
+  
+    var page = window.location.href;
+    sessionStorage.setItem("last_page", page);
+    
+    window.location.href = "playerProfile.html";
+  }
 }
 
 $(document).ready(() => {
@@ -46,38 +49,40 @@ $(document).ready(() => {
   }
 
   playerCount = 0;
-  parsedLastSearchedMatches.forEach(player => {
-    var playerWrapper = document.createElement("div");
-    playerWrapper.className = "last-searched-player";
-    playerWrapper.setAttribute('onclick', 'openSearchedPlayer(this.lastChild.textContent)');
-
-    var playerNameTag = document.createElement("span");
-    playerNameTag.className = "player-name-tag";
-    playerNameTag.textContent = player.player_name + "#" + player.player_tag;
-
-    var playerRegion = document.createElement("span");
-    playerRegion.className = "lastPlayer-player-region";
-    playerRegion.textContent = player.player_region.toUpperCase();
-
-    var deleteSearchedPlayer = document.createElement("i");
-    deleteSearchedPlayer.classList.add("fas", "fa-times", "removeLastSearch");
-    deleteSearchedPlayer.setAttribute('onclick', 'deleteSearchedPlayer(this.parentElement, this.id); event.stopPropagation();');
-    deleteSearchedPlayer.id = playerCount;
-
-    var searchedPlayerData = document.createElement("span");
-    searchedPlayerData.setAttribute('style', 'display: none;');
-    searchedPlayerData.textContent = player.puuid + ";" + player.player_name + ";" + player.player_tag + ";" + player.player_region + ";"
-
-    playerWrapper.appendChild(playerNameTag);
-    playerWrapper.appendChild(playerRegion);
-    playerWrapper.appendChild(deleteSearchedPlayer);
-    playerWrapper.appendChild(searchedPlayerData);
-
-    var wrapper = document.getElementById('last-searched-players');
-    var lastElement = document.getElementById('lastSPElement')
-    wrapper.insertBefore(playerWrapper, lastElement)
-    playerCount++;
-  });
+  if(fs.existsSync(process.env.APPDATA + '/VALTracker/user_data/player_search/last_searched_players.json')) {
+    parsedLastSearchedMatches.forEach(player => {
+      var playerWrapper = document.createElement("div");
+      playerWrapper.className = "last-searched-player";
+      playerWrapper.setAttribute('onclick', 'openSearchedPlayer(this.lastChild.textContent)');
+  
+      var playerNameTag = document.createElement("span");
+      playerNameTag.className = "player-name-tag";
+      playerNameTag.textContent = player.player_name + "#" + player.player_tag;
+  
+      var playerRegion = document.createElement("span");
+      playerRegion.className = "lastPlayer-player-region";
+      playerRegion.textContent = player.player_region.toUpperCase();
+  
+      var deleteSearchedPlayer = document.createElement("i");
+      deleteSearchedPlayer.classList.add("fas", "fa-times", "removeLastSearch");
+      deleteSearchedPlayer.setAttribute('onclick', 'deleteSearchedPlayer(this.parentElement, this.id); event.stopPropagation();');
+      deleteSearchedPlayer.id = playerCount;
+  
+      var searchedPlayerData = document.createElement("span");
+      searchedPlayerData.setAttribute('style', 'display: none;');
+      searchedPlayerData.textContent = player.puuid + ";" + player.player_name + ";" + player.player_tag + ";" + player.player_region + ";"
+  
+      playerWrapper.appendChild(playerNameTag);
+      playerWrapper.appendChild(playerRegion);
+      playerWrapper.appendChild(deleteSearchedPlayer);
+      playerWrapper.appendChild(searchedPlayerData);
+  
+      var wrapper = document.getElementById('last-searched-players');
+      var lastElement = document.getElementById('lastSPElement')
+      wrapper.insertBefore(playerWrapper, lastElement)
+      playerCount++;
+    });
+  }
 
   // Execute a function when the user releases a key on the keyboard
   $("#playerNameSearch").keyup(function (event) {
